- move MockFileSystemManager to separate file - made SyncFileTest class with unit-tests for SyncFile - implemented FileSystemManager with System.IO using - converted FileMetada.ByteSize field from int to long - redesigned SyncFile class logic and inner structure:   - removed fields ByteSize and LastModified, now just return FileMetadata object from RetrieveFileMetadata method   - removed RetrieveFileMetadata calls from constructors (cause it's possible exception throwing inside)

+++ BeaverSync/BeaverSyncLib/IFileSystemManager.cs
@@ -40,6 +40,6 @@ namespace BeaverSyncLib
         /// <summary>
         /// Размер файла в байтах
         /// </summary>
-        public int ByteSize { get; set; }
+        public long ByteSize { get; set; }
     }
 }
